requirements.txt
Django==4.2.7
Pillow==10.1.0
django-cloudinary-storage==0.3.0
cloudinary==1.36.0
python-decouple==3.8
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
django-countries==7.5.1
django-extensions==3.2.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0